refactor: フッター配置スタイルをflex式に変更、クローラー除外設定追加(#10)
